'Select Message' Improvement - More recent messages
Improved 'Select Message' function. The bot will now check if the chosen message is one of the 10 most recently posted instead of just the last post and pick another message if it is.

This is customisable, for example, I can make the bot check the 20 most recent or 5 most recent without having to change the code.